backend/requirements/base.txt
Django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.2
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
celery==5.3.1
redis==4.6.0
psycopg2-binary==2.9.6
Pillow==10.0.0
numpy==1.25.1
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0 